refactor: move logic to LLC and fix mobile injected code
